+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.6](https://github.com/realityforge/gwt-keycloak/tree/v0.6) (2020-02-20) · [Full Changelog](https://github.com/realityforge/gwt-keycloak/compare/v0.5...v0.6)
 
 Changes in this release: